application/requirements.txt
flask
snowflake-connector-python
python-dotenv
pandas
